Rename big_tree benchmarks to big_tree_flexbox
